requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Market data
yfinance>=0.2.0

# AI
anthropic>=0.18.0

# Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Scheduling (optional)
apscheduler>=3.10.0
